pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "bioverse"
dynamic = ["version"]
description = "A standardized framework for machine learning on biomolecules."
readme = "README.md"
license = {text = "BSD-3-Clause"}
authors = [
    {name = "Tim Kucera"}
]
maintainers = [
    {name = "Tim Kucera"}
]
keywords = ["bioinformatics", "machine-learning", "proteins", "rna", "molecular-biology"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "awkward>=1.8.0",
    "biopandas>=0.2.7",
    "rich>=12.0.0",
    "requests>=2.25.0",
    "joblib>=1.1.0",
    "pyarrow>=7.0.0",
    "multiprocess>=0.70.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "coverage>=6.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]
docs = [
    "sphinx>=4.0.0",
    "sphinxawesome-theme==5.3.2",
]

[project.urls]
Homepage = "https://github.com/BorgwardtLab/bioverse"
Documentation = "https://bvrs.readthedocs.io/"
Repository = "https://github.com/BorgwardtLab/bioverse"
"Bug Tracker" = "https://github.com/BorgwardtLab/bioverse/issues"

[project.scripts]
bioverse = "bioverse.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["bioverse*"]

[tool.setuptools.package-data]
bioverse = ["*.txt", "*.json", "*.yaml", "*.yml"]

[tool.setuptools_scm]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["bioverse"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["testing"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["bioverse"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/venv/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
] 